auth/requirements.txt
alembic==1.9.0
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
